requirement.txt
python-magic
tkintertable